+++ src/__generated__/SaleArtworksGridQuery.graphql.ts
@@ -59,6 +59,8 @@ fragment Artwork_artwork on Artwork {
   date
   sale_message
   is_in_auction
+  is_biddable
+  is_acquireable
   id
   sale_artwork {
     opening_bid {
@@ -157,7 +159,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "SaleArtworksGridQuery",
-  "id": "fe2fcd91244ec95555407491413ecb12",
+  "id": "59517238e770fe265224ee1deec8e590",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -309,7 +311,7 @@ return {
                               {
                                 "kind": "ScalarField",
                                 "alias": null,
-                                "name": "sale_message",
+                                "name": "is_in_auction",
                                 "args": null,
                                 "storageKey": null
                               },
@@ -366,11 +368,25 @@ return {
                                 "args": null,
                                 "storageKey": null
                               },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "sale_message",
+                                "args": null,
+                                "storageKey": null
+                              },
                               v2,
                               {
                                 "kind": "ScalarField",
                                 "alias": null,
-                                "name": "is_in_auction",
+                                "name": "is_biddable",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "is_acquireable",
                                 "args": null,
                                 "storageKey": null
                               },