+++ Example/Emission/EigenLikeAdminViewController.h
@@ -3,7 +3,7 @@
 
 /// Tickable cells
 extern NSString *const AROptionCell;
-/// Tickable cells with a subtitle
+/// Cells with a subtitle
 extern NSString *const ARSubtitledLabOptionCell;
 /// Cells with a preload button
 extern NSString *const ARPreloadOptionCell;