+++ src/__generated__/WorksForYouQuery.graphql.ts
@@ -92,6 +92,8 @@ fragment Artwork_artwork on Artwork {
   date
   sale_message
   is_in_auction
+  is_biddable
+  is_acquireable
   id
   sale_artwork {
     opening_bid {
@@ -149,7 +151,7 @@ v1 = {
 v2 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "date",
+  "name": "sale_message",
   "args": null,
   "storageKey": null
 },
@@ -224,7 +226,7 @@ v8 = {
 v9 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "sale_message",
+  "name": "date",
   "args": null,
   "storageKey": null
 },
@@ -235,7 +237,21 @@ v10 = {
   "args": null,
   "storageKey": null
 },
-v11 = [
+v11 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "is_biddable",
+  "args": null,
+  "storageKey": null
+},
+v12 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "is_acquireable",
+  "args": null,
+  "storageKey": null
+},
+v13 = [
   {
     "kind": "ScalarField",
     "alias": null,
@@ -244,7 +260,7 @@ v11 = [
     "storageKey": null
   }
 ],
-v12 = {
+v14 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "sale_artwork",
@@ -261,7 +277,7 @@ v12 = {
       "args": null,
       "concreteType": "SaleArtworkOpeningBid",
       "plural": false,
-      "selections": v11
+      "selections": v13
     },
     {
       "kind": "LinkedField",
@@ -271,7 +287,7 @@ v12 = {
       "args": null,
       "concreteType": "SaleArtworkCurrentBid",
       "plural": false,
-      "selections": v11
+      "selections": v13
     },
     {
       "kind": "ScalarField",
@@ -302,11 +318,11 @@ v12 = {
     v1
   ]
 },
-v13 = [
+v15 = [
   v5,
   v1
 ],
-v14 = {
+v16 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "partner",
@@ -314,9 +330,9 @@ v14 = {
   "args": null,
   "concreteType": "Partner",
   "plural": false,
-  "selections": v13
+  "selections": v15
 },
-v15 = {
+v17 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "image",
@@ -362,7 +378,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "WorksForYouQuery",
-  "id": "03402e06719b105d2d1174c8d3b0159a",
+  "id": "859cfe2b573b62fd12ad8880be6e34c9",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -548,15 +564,17 @@ return {
                                   v6,
                                   v7,
                                   v8,
-                                  v1,
                                   v9,
+                                  v1,
                                   v10,
+                                  v11,
                                   v12,
                                   v14,
+                                  v16,
                                   v4
                                 ]
                               },
-                              v15,
+                              v17,
                               {
                                 "kind": "ScalarField",
                                 "alias": null,
@@ -631,7 +649,7 @@ return {
               v6,
               v4,
               v5,
-              v15,
+              v17,
               {
                 "kind": "LinkedField",
                 "alias": null,
@@ -654,14 +672,16 @@ return {
                 "concreteType": "Artwork",
                 "plural": true,
                 "selections": [
-                  v9,
+                  v10,
                   v1,
                   v7,
                   v8,
+                  v9,
                   v2,
                   v6,
-                  v10,
+                  v11,
                   v12,
+                  v14,
                   {
                     "kind": "LinkedField",
                     "alias": null,
@@ -670,9 +690,9 @@ return {
                     "args": v3,
                     "concreteType": "Artist",
                     "plural": true,
-                    "selections": v13
+                    "selections": v15
                   },
-                  v14,
+                  v16,
                   v4
                 ]
               },