+++ src/lib/Components/ArtworkGrids/Artwork.tsx
@@ -1,14 +1,29 @@
+import { Sans } from "@artsy/palette"
+import colors from "lib/data/colors"
+import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { map } from "lodash"
 import React from "react"
 import { Image, StyleSheet, TouchableWithoutFeedback, View } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
-
-import colors from "lib/data/colors"
-import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import styled from "styled-components/native"
 import ImageView from "../OpaqueImageView"
 import SerifText from "../Text/Serif"
 
 import { Artwork_artwork } from "__generated__/Artwork_artwork.graphql"
+
+const Badges = styled.View`
+  position: absolute;
+  bottom: 16px;
+  left: 1px;
+  display: flex;
+  flex-direction: row;
+`
+const Badge = styled.View`
+  border-radius: 2px;
+  padding: 3px 5px 1px 5px;
+  background-color: white;
+  margin-left: 5px;
+`
 
 interface Props {
   artwork: Artwork_artwork
@@ -34,13 +49,46 @@ class Artwork extends React.Component<Props, any> {
     return (
       <TouchableWithoutFeedback onPress={this.handleTap.bind(this)}>
         <View>
-          <ImageView style={styles.image} aspectRatio={artwork.image.aspect_ratio} imageURL={artwork.image.url} />
+          <View>
+            <ImageView style={styles.image} aspectRatio={artwork.image.aspect_ratio} imageURL={artwork.image.url} />
+            {this.badges()}
+          </View>
           {this.artists()}
           {this.artworkTitle()}
           {partnerName && <SerifText style={styles.text}>{partnerName}</SerifText>}
           {this.saleMessage()}
         </View>
       </TouchableWithoutFeedback>
+    )
+  }
+
+  hasBadges() {
+    const { is_acquireable, is_biddable } = this.props.artwork
+    return is_acquireable || is_biddable
+  }
+
+  badges() {
+    const { is_acquireable, is_biddable } = this.props.artwork
+
+    return (
+      this.hasBadges() && (
+        <Badges>
+          {is_acquireable && (
+            <Badge>
+              <Sans weight="medium" size="1">
+                BUY NOW
+              </Sans>
+            </Badge>
+          )}
+          {is_biddable && (
+            <Badge>
+              <Sans weight="medium" size="1">
+                BID
+              </Sans>
+            </Badge>
+          )}
+        </Badges>
+      )
     )
   }
 
@@ -121,6 +169,8 @@ export default createFragmentContainer(
       date
       sale_message
       is_in_auction
+      is_biddable
+      is_acquireable
       id
       sale_artwork {
         opening_bid {