+++ CHANGELOG.md
@@ -16,10 +16,10 @@
 
 ### Master
 
-* Adding artwork badges to brick - matt
 ### 1.6.0
 
 * Add @artsy/palette - yuki/orta/chris
+* Adding artwork badges to brick - matt
 
 ### 1.5.15