+++ CHANGELOG.md
@@ -18,8 +18,8 @@
 
 ### 1.6.0
 
-* Add @artsy/palette - yuki/orta/chris
 * Adding artwork badges to brick - matt
+* Add @artsy/palette - yuki/orta/chris
 
 ### 1.5.15