+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### 1.6.0
 
+* Adding artwork badges to brick - matt
 * Add @artsy/palette - yuki/orta/chris
 
 ### 1.5.15